Add specific loadImplementation in register and shim
- lib/index.js accepts root object and loadImplementation function
- Node.js loadImplementation contains logic for auto detect and global
  Promise based on Node.js version. Uses global as root.
- Browser specific loadImplementation always loads window.Promise. in
  register-shim.js. Uses window as root.
