+++ register-shim.js
@@ -1,2 +1,9 @@
 "use strict";
-module.exports = require('./lib')(true);
+module.exports = require('./lib')(window, loadImplementation)
+
+function loadImplementation(){
+  return {
+    Promise: window.Promise,
+    implementation: 'window.Promise'
+  }
+}
